practice for CCC 2021
Deciding to compete in CCC junior 2021, I began to do some previous years' questions for practice. Not all of them work and some of them have jank solutions but overall it was good practice. I'm especially proud of my breadth-first search for CCC 2020 J5. If you're wondering why the CCC is lowercase for the 2021 code, those were the requirements of the competition and I never changed it after moving it to this repository